requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
sqlalchemy==2.0.29
asyncpg==0.29.0